+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.1 Under development
 
+- Enh #20: Add `normalizeKey()` method in `BaseAttribute` helper for normalization attribute keys with prefixes, enhance error handling and add related tests (@terabytesoftw)
+
 ## 0.5.0 December 19, 2025
 
 - Bug #15: Update copyright year to `2024` in `LICENSE` file (@terabytesoftw)
@@ -9,7 +11,6 @@
 - Bug #17: Enhance support for `Stringable` interface types across `BaseCssClass`, `BaseEncode`, `BaseEnum`, `BaseValidator` classes and update related tests (@terabytesoftw)
 - Bug #18: Update feature descriptions in SVG files to reflect new capabilities including `Stringable` support and performance optimizations (@terabytesoftw)
 - Bug #19: Remove trailing periods from alert content in SVG feature descriptions (@terabytesoftw)
-- Enh #20: Add `normalizeKey()` method in `BaseAttribute` helper for normalization attribute keys with prefixes, enhance error handling and add related tests (@terabytesoftw)
 
 ## 0.4.0 December 17, 2025