+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.5.1 Under development
 
-- Enh #20: Add `normalizeKey()` method in `BaseAttribute` helper for normalization attribute keys with prefixes, enhance error handling and add related tests (@terabytesoftw)
+- Enh #20: Add `normalizeKey()` method in `BaseAttributes` helper for normalization attribute keys with prefixes, enhance error handling and add related tests (@terabytesoftw)
 
 ## 0.5.0 December 19, 2025